feat: add listeners to changes in currently opened files
This commit modifies the `startSession` function in `commands.ts` to include the `context` parameter. It then uses the `vscode.workspace.textDocuments` API to retrieve a list of currently opened files and adds them to the `ymap` object. Event listeners are added to track changes in opened files and update the `ymap` accordingly.

TODO: add disposables to `context`